pyproject.toml
[tool.poetry]
name = "my-package"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.3"
factory-boy = "^3.2.0"
django-extensions = "^3.1.3"
pytest-django = "^4.3.0"
pytest-cov = "^2.12.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.8.0"
pycodestyle = "^2.7.0"
ipdb = "^0.13.7"
pytest = "^6.2.4"
pytest-django = "^4.3.0"
pytest-cov = "^2.12.0"

[tool.black]
line-length = 79
exclude = """
/(
    | statistics
    | media
    | templates
    | core.migrations
    | core.templates
)/
"""

[tool.isort]
multi_line_output = 3
profile = 'black'
[build-system]

requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

